+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 
 ## Future
 - Added TRIO109: Async definitions should not have a `timeout` parameter. Use `trio.[fail/move_on]_[at/after]`
-- Added TRIO301: `while <condition>: await trio.sleep()` should be replaced by a `trio.Event`.
+- Added TRIO110: `while <condition>: await trio.sleep()` should be replaced by a `trio.Event`.
 
 ## 22.7.6
 - Extend TRIO102 to also check inside `except BaseException` and `except trio.Cancelled`